+++ crates/sdk/src/js/price.rs
@@ -1,15 +1,18 @@
 use serde::{Deserialize, Serialize};
 use tsify_next::Tsify;
 
+use crate::market::Value;
+
 /// Js Prices.
 #[derive(Debug, Serialize, Deserialize, Tsify)]
+#[tsify(into_wasm_abi, from_wasm_abi)]
 pub struct Prices {
     /// Index token price.
-    pub index_token: Price,
+    pub index_token: Value,
     /// Long token price.
-    pub long_token: Price,
+    pub long_token: Value,
     /// Short token price.
-    pub short_token: Price,
+    pub short_token: Value,
 }
 
 impl From<Prices> for gmsol_model::price::Prices<u128> {
@@ -18,25 +21,6 @@ impl From<Prices> for gmsol_model::price::Prices<u128> {
             index_token_price: value.index_token.into(),
             long_token_price: value.long_token.into(),
             short_token_price: value.short_token.into(),
-        }
-    }
-}
-
-/// Js Price.
-#[derive(Debug, Serialize, Deserialize, Tsify)]
-#[tsify(from_wasm_abi)]
-pub struct Price {
-    /// Min price.
-    pub min: u128,
-    /// Max price.
-    pub max: u128,
-}
-
-impl From<Price> for gmsol_model::price::Price<u128> {
-    fn from(value: Price) -> Self {
-        Self {
-            min: value.min,
-            max: value.max,
         }
     }
 }